+++ src/main/java/com/group07/PetHealthCare/service/VeterinarianService.java
@@ -1,5 +1,8 @@
 package com.group07.PetHealthCare.service;
 
+import com.group07.PetHealthCare.exception.AppException;
+import com.group07.PetHealthCare.exception.ErrorCode;
+import com.group07.PetHealthCare.pojo.Customer;
 import com.group07.PetHealthCare.pojo.Veterinarian;
 import com.group07.PetHealthCare.respositytory.VeterinarianRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,10 +20,18 @@ public class VeterinarianService {
 
     public Optional<Veterinarian> getVeterinarianById (String Id)
     {
+        Optional<Veterinarian> veterinarian = veterinarianRepository.findById(Id);
+        if(!veterinarian.isPresent()){
+            throw new AppException(ErrorCode.NOT_FOUND);
+        }
         return veterinarianRepository.findById(Id);
     }
     public void deleteVeterinarianById(String Id)
     {
+        Optional<Veterinarian> veterinarian = veterinarianRepository.findById(Id);
+        if(!veterinarian.isPresent()){
+            throw new AppException(ErrorCode.NOT_FOUND);
+        }
         veterinarianRepository.deleteById(Id);
     }
